middleware: migrate scheduling tests to test_schedule_workbook.py — Issue #98 Step 4
Move pure scheduling-method tests off ChurchTeamsExporter and onto
ScheduleWorkbookBuilder. Tests that exercise _write_excel_report stay in
test_church_teams_export.py since they cover live-export integration.
